Move tt commands to commands/time_trial
