+++ commands/time_trial/delete_submission.ts
@@ -1,7 +1,6 @@
 import { CacheType, ChatInputCommandInteraction, EmbedBuilder, SlashCommandBuilder } from "discord.js";
-import { getConfig } from "../config.js";
-import { resolveModRestrictPermission } from "../utils.js";
-import { handleProfileAutocomplete } from "../tt-utils.js";
+import { getConfig } from "../../config.js";
+import { resolveModRestrictPermission } from "../../utils.js";
 
 const config = getConfig();
 
@@ -10,30 +9,22 @@ export default {
     adminOnly: false,
 
     data: new SlashCommandBuilder()
-        .setName("tt_profile_delete")
-        .setDescription("Delete a Time Trial profile (must have no submissions)")
-        .addStringOption(option => option
-            .setName("profile")
-            .setDescription("Profile to delete")
-            .setRequired(true)
-            .setAutocomplete(true))
+        .setName("tt_delete_submission")
+        .setDescription("Delete a Time Trial ghost submission")
+        .addIntegerOption(option => option
+            .setName("submission_id")
+            .setDescription("The submission ID to delete")
+            .setRequired(true))
         .setDefaultMemberPermissions(resolveModRestrictPermission()),
 
-    autocomplete: async function(interaction: any) {
-        const focusedOption = interaction.options.getFocused(true);
-        if (focusedOption.name === "profile") {
-            await handleProfileAutocomplete(interaction);
-        }
-    },
-
     exec: async function(interaction: ChatInputCommandInteraction<CacheType>) {
-        const profileId = interaction.options.getString("profile", true);
+        const submissionId = interaction.options.getInteger("submission_id", true);
 
         await interaction.deferReply();
 
         const leaderboardUrl = `http://${config.leaderboardServer}:${config.leaderboardPort}`;
         try {
-            const response = await fetch(`${leaderboardUrl}/api/moderation/timetrial/profile/${profileId}`, {
+            const response = await fetch(`${leaderboardUrl}/api/moderation/timetrial/submission/${submissionId}`, {
                 method: "DELETE",
                 headers: { "Authorization": `Bearer ${config.wfcSecret}` }
             });
@@ -43,21 +34,24 @@ export default {
 
                 const embed = new EmbedBuilder()
                     .setColor(0xff0000)
-                    .setTitle("🗑️ TT Profile Deleted")
+                    .setTitle("🗑️ Ghost Submission Deleted")
                     .setDescription(result.message)
+                    .addFields(
+                        { name: "Submission ID", value: submissionId.toString(), inline: true }
+                    )
                     .setTimestamp();
 
                 await interaction.editReply({ embeds: [embed] });
             } else {
                 const errorData = await response.json();
                 await interaction.editReply({
-                    content: `Failed to delete profile: ${errorData.message || response.statusText}`
+                    content: `Failed to delete submission: ${errorData.message || response.statusText}`
                 });
             }
         } catch (error) {
-            console.error("Error deleting TT profile:", error);
+            console.error("Error deleting submission:", error);
             await interaction.editReply({
-                content: "Network error while deleting profile"
+                content: "Network error while deleting submission"
             });
         }
     }
