+++ commands/time_trial/profile_delete.ts
@@ -1,6 +1,7 @@
 import { CacheType, ChatInputCommandInteraction, EmbedBuilder, SlashCommandBuilder } from "discord.js";
-import { getConfig } from "../config.js";
-import { resolveModRestrictPermission } from "../utils.js";
+import { getConfig } from "../../config.js";
+import { resolveModRestrictPermission } from "../../utils.js";
+import { handleProfileAutocomplete } from "./tt_utils.js";
 
 const config = getConfig();
 
@@ -9,22 +10,30 @@ export default {
     adminOnly: false,
 
     data: new SlashCommandBuilder()
-        .setName("tt_delete_submission")
-        .setDescription("Delete a Time Trial ghost submission")
-        .addIntegerOption(option => option
-            .setName("submission_id")
-            .setDescription("The submission ID to delete")
-            .setRequired(true))
+        .setName("tt_profile_delete")
+        .setDescription("Delete a Time Trial profile (must have no submissions)")
+        .addStringOption(option => option
+            .setName("profile")
+            .setDescription("Profile to delete")
+            .setRequired(true)
+            .setAutocomplete(true))
         .setDefaultMemberPermissions(resolveModRestrictPermission()),
 
+    autocomplete: async function(interaction: any) {
+        const focusedOption = interaction.options.getFocused(true);
+        if (focusedOption.name === "profile") {
+            await handleProfileAutocomplete(interaction);
+        }
+    },
+
     exec: async function(interaction: ChatInputCommandInteraction<CacheType>) {
-        const submissionId = interaction.options.getInteger("submission_id", true);
+        const profileId = interaction.options.getString("profile", true);
 
         await interaction.deferReply();
 
         const leaderboardUrl = `http://${config.leaderboardServer}:${config.leaderboardPort}`;
         try {
-            const response = await fetch(`${leaderboardUrl}/api/moderation/timetrial/submission/${submissionId}`, {
+            const response = await fetch(`${leaderboardUrl}/api/moderation/timetrial/profile/${profileId}`, {
                 method: "DELETE",
                 headers: { "Authorization": `Bearer ${config.wfcSecret}` }
             });
@@ -34,24 +43,21 @@ export default {
 
                 const embed = new EmbedBuilder()
                     .setColor(0xff0000)
-                    .setTitle("🗑️ Ghost Submission Deleted")
+                    .setTitle("🗑️ TT Profile Deleted")
                     .setDescription(result.message)
-                    .addFields(
-                        { name: "Submission ID", value: submissionId.toString(), inline: true }
-                    )
                     .setTimestamp();
 
                 await interaction.editReply({ embeds: [embed] });
             } else {
                 const errorData = await response.json();
                 await interaction.editReply({
-                    content: `Failed to delete submission: ${errorData.message || response.statusText}`
+                    content: `Failed to delete profile: ${errorData.message || response.statusText}`
                 });
             }
         } catch (error) {
-            console.error("Error deleting submission:", error);
+            console.error("Error deleting TT profile:", error);
             await interaction.editReply({
-                content: "Network error while deleting submission"
+                content: "Network error while deleting profile"
             });
         }
     }
